Move server/config.go into its own package
